requirements.txt
discord.py
numpy
matplotlib
python-dotenv
unidecode